Restructure files into modules
Closes #4
